Fix gpt-4o-audio models on Termux and `options -vv` (low latency of response).
---

KNOWN OR PRESUMED BUGS: token count of the audio-models kind of works but not perfectly. It is also hard to try and decide what to record, because audio models have reasoning tokens and audio tokens, and audio tokens may be present for the user input and/or as part of the output!

---

TO-DO for next releases:
  #add test for filename with white spaces separated by pipe in test suite.
  #consider checking simply $MULTIMODAL>0, instead of $MULTIMODAL>1.
  #reconsider the additional audio-model related defaults $INSTRUCTION_CHAT text appended when all configs are defaults. Simplify it?

---