+ Small maintenance update.
Renamed function _mediachatf() to media_pathf(), because the hyphen before the function name did not really align with the other function names and conventions used in this script.

Timeout user-required key press after printing the command output to the terminal, and the user is using the text editor mode, after 6 seconds. I reckon that should be enough for this small case.

Wait 1.6 seconds after an error response is printed to stderr if the user prompt is bigger than 1640 charactters. This is necessary to give the user some time to check the error message before the script goes into retry mode, and add the last (big) prompt is loaded into readline or text editor again, thus pushing the error message much further up the terminal scroll buffer!...

---