New feat: Novita AI integration, various bug fixes
We have integrated the Novita AI API code. [Check the docs](https://github.com/mountaineerbr/shellChatGPT#novita-ai).

Update know model list to set maximum capacity automatically. Update known model costs.

Make changes in the code in order to improve API service provider integration.
Should be backward compatible with the old code, but the logic is not perfected.

Improve regex in the processing pipeline of streaming response.

Add new feature: at the user confirmation prompts, if the user press "-" or "/" or "!", he will be taken to a new command prompt to run a chat command before being asked again about confirmation of the input again. This feature may not work with all the chat commands, but it should be useful to set the model configuration at the last second before the request (without disrupting current input buffer). This is why this feature is not going to be documented any time soon, as we would need much more testing to be sure of its limitations.. But it is working as expected..

The youtube video transcription dump function will also try and dump the video description.

---