Fix Audio-Model code under Termux, improved interface when using audio-models.
Under Termux, set the right software to work with audio-models voice recording system.
Sox or ffmpeg are required.

Adjust voice recording file input format.

Adjust some code related to the audio-model interface, and also in the main prompt requesting function.

Give more time to the user to choose if he wants to replay the audio file that we received from audio-models.

Make sure to print the error response in audio-model use cases (so the user can work out the problems if he is doing something that is not compatible with the API).

---