Fix regressions related to recent code refactoring related to Mistral AI and the new Novita AI.
The regressions prevented using Whisper from OPENAI when setitng Novita AI and Mistral AI integrations. So now the `-ccwz --mistral` command line restores how the script wraps Mistral/Novita AI and fall back to OpenAI for Whisper and TTS as expected.

Also, change the hallmark X on curl request to C. This is because the user can cancel the request with CTRL+C, so I believe that printing out a C is better as a user interface to tip the user of CTRCL+C usage.....

---