Change command line "option -d" to single-turn text completions, and "-dd" for multi-turn text completions.
Since changing the defaults demo mode (when no options are set by the user whatsoever) to point to chat completions in single-turn, we lost an intuitive way to use the text completions mode in single-turn. The man page stated we should use options "-dE" combined to achieve that effect but that is a little hard to memorise, so I went ahead and duplicated option -d, so that a single -d command line option activates the text completions single-turn mode, and the double -dd activates the multi-turn mode of text completions with history support!

Also, make sure cURL only tests input that have a tld structure (is_linkf() function), also decrease cURL timeout for such cases...

---