Important old regression fix, and improve lax user interface with some shortcuts.
Important fix in the session_copyf() function. It would not resume anymore after copying the session since we changed how we deal with session management. Unfortunately, this regression bug is old, probably since the start of the year.

This regression dates back to 3 July 2024 at least, and probably dates back to 16 May 2024. I shall not test older script versions as that is just time wasted.

It is **now fixed**.

If the first positional argument is a dot ".", then enter interactive mode to determine which old session we should resume from. Before, the shortcut was "/." or the command "/sub", but these are hardly intuitive!

Also, typing a forward slash "/" as the first positional argument opens the user custom prompt selector.

Improve setting whisper + chat, and chat + tts option parsing from the command line. We do some little divination to guess if the user is trying to set options to whisper or tts, or trying to set the first user prompt from the command line. In the voice-in and voice-out of chat, the proper syntax to set options from the command line would be:

   chatgpt.sh -ccwz [whisper_opts] -- [tts_opts] -- [text_prompt]

However, this is not an easy syntax to convey to the user, or for him to remember. I reckon in most cases, the user will set options -ccwz and then maybe try to input the first user prompt from the command line (even when whisper is set for posterior turn voice input!).

So we try to determine whether the user means to set whisper/tts options or to set a text input as the first user prompt **based on user input length**.

Add command line options --awesome and --awesome-zh as aliases to -S/ and -S%. This should make using awesome prompts easier than the single letter option and operators.

Update help, man and readme pages.

---