Image Variations: when setting the alpha colour and fuzz, make sute to edit the image with ImageMagick.
The code to detect if the input file had a transperent alpha channel has been working just fine.

However, from a user point of view, when setting the alpha colour option -@,
 such as `chatgpt.sh -i -@ white8% ~/twhite.png "A sky with eagles."`,
the script now makes sure to edit the alpha colour even if the input image has a transparent alpha channel already!

Also, in chat mode, when running `-1.2` it is now alised to `/temp 1.2`. This avoids messing the model input and capacity settings when inadvertidely forggeting to run the above alias as the correct `-t1.2` command in chat mode...

---
```
Improve alpha color handling for transparent images and alias `/temp` for chat mode temperature inputs.
```