Anthropic `claude-3-7-sonnet` integration and bug fixes.
Set claude-3-7-sonnet-latest as the defaults model for Anthropic.

---

Implement the command line --think (or --effort) command line option to set the maximum thinking tokens for the claude-3-7 models.

**Do mind that** that to activate the thinking process, the user must specify either `--think [NUM]` (or `--effort [NUM]`) command line option. To activate thinking during chat, use `!think`, `/think`, or `/effort` commands.

There is no way to make a clear display separation of the thinking process and the actual answer, which is unfortunate because it becomes hard to know what is thinking and what is answer as they will be printed out without any visible  separation!

This new code may still need some improvements overall...

---

Do not record the thinking / reasoning process to the history file.
Thinking is really just a model internal thinggie and we should not bother to keep the thinking process on record (this avoids sending unnecessary thinking process text to the model on new turns).

---

Fix the `/jump` and `//jump` chat commands. This has been broken for a long time, since we changed how to script works with respect to recording instructions, user and assistant messages.

The `/jump` and `//jump` chat commands relied on the old mechanism of recording to the history file right away and then reading those entries.
Since we changed when and how the script records the entries, these commands were pretty much broken.

Probably one of the last bugs related to the "new" year-old history file recording system.

---

Improve the `/url` command to dump webpage text. Use curl in case w3m/links/elinks/lynx cannot dump the text. This may happen sometimes. So we set a special crafted curl command with usually works in these scenarios, nonetheless.

---

Update docs and the config file template.

---