Add support for `gpt-4o-tts` model series (Text-To-Speech), and for gpt-4o-transcribe (Transriptions).
Defaults TTS to gpt-4o-tts for speech synthesis.

NEW ENVAR: `$INSTRUCTION_SPEECH` may be set to instruct the audio model to speak in specific ways. This envar is **only** used with the TTS stand-alone `option -z` on script invocation.

Note that to set the tone and accent of the rendered voice output by audio models (gpt-4o-audio), set a robust INSTRUCTION as usual.

We still use `whisper-1` as our defaults for audio/voice transcription on because this is a cheaper model and works well for most use cases with the OpenAI API.

Got some help from OpenAI community this time.

<https://community.openai.com/t/request-to-gpt-4o-mini-transcribe-model/1151090/2>

---