Enable auto detection of OpenAI o-models as vision enabled models.
Since launch of o1-model proper (not the preview version which is not vision enabled), I had not updated the script to add o-models auto detection as vision models.  In these cases, the user may invoke the script with the --vision flag which sets the models as multimodal manually, anyways.

---