Set gpt-image-1 as default image generation and editing model.
New default model for image generation and editing as "gpt-image-1",
and update code to also support gpt-image new sizes and image quality,
even though the syntax of user interface is not optimal, it
remains the same as we are used to for dall-e-3.

Mind that for using gpt-imge models, the user must verify his organisation
on OpenAI website! Dalle-e-3 and dall-e-2 compatibility is maintained!

Add some more OpenAI models to the cost calculation function. We may remove the
the cost calculation at some point, as this demands pro-active maintenance
in adding new models and their costs to the script source code case function,
which is a boring and eternal task.

Revert a recent update in which we tested if the model was an audio model
before testing it as a vision model, because there are few audio models
and far more vision models available. Keep test order audio->vision only
when strictly necessary!  This may improve performance even though only
ever slightly.

Print error response on image generation and editing endpoints.

If user adds an image or audio with the chat command "/media [path]",
it automatically sets the model as multimodal.

Some small performance improvements, specially in the media_pathf() and set_histf()
functions, change "[[" tests to "case" glob tests.
The media_pathf() performance gain is about ~3.6% when multiple image input
is processed, but this result came from very limited testing.

Update default models of some supported platforms.

Update docs and man pages.

---