Implemente "native" web search in OpenAI, Gemini, Anthropic AI, and xAI API requests.
When use runs the chat command `/g [prompt]` in the above service  proveiders,
the apporpriate code is added to the requets body to turn on their native web search
dump tool.

If the user wants to use the fallback and in-house solution for these APIs nonetheless,
he can run  `//g [prompt]`. This will ask the user to choose from Google, DuckDuckGo, and Brave.
This usually qorks better with *w3m* or other cli-browsers supported by the script.

Add support for Anthropic "claude-opus-4" model.

Fix recording the the amount of tokens used in a chat turn, and token rate of the Groq API.

Fix  GitHub Model catalog, options -l and -ll, with an optional model name for details.

Do not include parameter "n" (number of results) in the text/chat completions when this is "1".

Do not include temperature parameter when this is set to empty/null with command line option "--temperature ''".

Add support for xAI image generation model grok-2-image-1212.

Some general bug fixes and code maintenance in Bash code.

Update some default models of some providers.

Update docs and readme.md.

---