Code maintenance and bug fixes of secondary/tertiary features. Performance improvements!
The shell functions that were used to glob-trim input in many instances were updated to use new input-trimming functions which are faster.

The new trimming functions do not need to be run as forks but do require more code to set the variables whose input is being glob-trimmed.

This required updating all instances where trimming functions were used, but and importantly, should not affect the way the script works at all! However, the code became even more difficult to work on and delicate for future maintenance, but as this script is but for myself, I don't really care..

A few tests revealed a 26% speed improvement in how the chat history is read from the record file and set for a new API request as this happens every turn fo the conversation, it should have the most noticeable speed improvement mechanism. It was not too slow before either, which makes me wonder if  this performance improvement update was really necessary (hopefully I don't break other things with this)!

---

Other than that, code maintenance was made to update how some script-specific features are implemented. Some of these lesser-used features of the script became degraded because of the many code changes that happened last year and really needed maintenance to work properly again. Hopefully these lesser-used script features are now working back again the way they are intended.

Unfortunately, very few people use this script regularly, if anyone at all, and probably they do not use the lesser-known features and do not raise isssues at the project repos, so I can only fix these bugs when I notice them or need to review big parts of the script code as I did today to improve the script performance.

---