Improve auto-set capacity of known models. Code maintenance. Remove session/turn cost.
Removing the cost / price calculator as this feature demands too much maintenance as costs change constantly.

We prefer *to try* and maintain the auto-setting of model maximum capacity of supported providers and  models, even though that is not guaranteed (too much and constant work). Anyways, it is very easy to set the model capacity as per the manual page instructions.

---