Use `yt-dlp` to dump youtube subtitles/captions.
The old method we used  to use, (youtube api), seems not to be working anymore.

The old code still remains but if the YouTube API really does not respond our requests any longer for some more time, the old code shall be removed and we shall depend on yt-dlp.

Vimeo caption dumps seem to not work anymore and the caption dump code is also flagged as deprecated for future removal.

---