Some bug fixes related to the rehab'ed user buffering system and other fixes.
Improve the code that handles empty input for set_maxtknf().

Fix when asking user for confirmation when there is input in the prompt buffer,
related to buffering prompts when the user starts a prompt with ':'.

---

Find a bug in single-turn text completions -d option
#bug: input whitespace trimming is not blocked with single -d!
#note: text completions is deprecated by OpenAI

The command line option single -d cannot be handled very well without new code overhead.
This stems from how the script was first designed to handle the many modes of text or chat completions,
and text completions was a thing at the time. Text completions mode was favoured at the beginning,
and the chat mode of text completions was specifically favoured.

Then things changed and native chat completions became the standard mode
but still try to support all the text completions modes.

The exception is the single -d option for single turn plain text completions,
which should not run whitespace trimming and would add newlines with command ':' inputs.
We would like input to be be free of trimming or adding spaces in this mode.

So the solution is to use the -dd flag which puts the script in
plain text completions and avoids all trimming and space insertions.

A workaround is to force -dd to be single-turn setting command line option -ddEE.
Options -EE skip history file recording.

The workaround implemented at t his time disables readline
history for consistency with what would be otherwise expected.
This will avoid problems when bash readline history is causing problems.

---