Fix bug in function is_txturl when input is  avery long line with a url/text/image file at the very end.
Relax some test cases for longer  input, so we hope that these adjustments will deal with input that is long enough.

These bugs have  a little to do with the optimisations we are doing with bash code for it to run faster. But that costs us time and bug fixes because we usually restrict some features of bash to make it run our routines faster, and such feature cuts may cause bugs with unexpected input types.

---