Bug fixes for the text editor interface and minor standardisation of user interface logic.
----

Fix a bug with the text editor interface when trying to run chat commands such as `/cat: [path].txt` or `/url: [url]`.

Improve how we handle the prompt buffer.  Now if the user wants to clear the buffer, he must not confirm when asked for conformation.

In the text editor mode, the empy reply will follow two cofirmation questions, and when the prpompt is empty and the answer is "no" to the conformations, we clear the prompt buffer and the current prompt to start anew.

Pressing to redo the prompt will only cler the current prompt buffer and not the stacked one.

----

When the user runs a command to regenerate the last question buffer with editing before the new request, and instead submit an empty prompt, we now understand that the user does not want to proceed with the answer regeneration and revert commenting on the last lines of the hstory file.

----

Revert an old change in code which is probably now unecessary and is simpler.

----

Don't grab yt-dlp &1 output at all because it will dump subtitles to a file, so 1>&2.

----

some minor code maintenance and whitespace removal.

---