Code maintenance and small fixes.
Update help and man pages for accurate wording after the prompt buffering system. specially try to use 'prepend' when appropriate and possible instead of 'insert'.

Rename oddly-named $RINSERT var to $PREPEND to mean what it does.

Fix session_listf() to run in a forked shell as the session_globf() executes a cd command.

Add any `!CMD` not matching a chat command is executed by the shell
as a shortcut for `!sh CMD`.

Do not abort and undo the history file commenting of the two last entries in the `!!regenerate command` to have the answer regenerated when the user presses enter in a new empty prompt after editing (deleting) the last prompt.
Instead, just print a reminder to the user it is regenerate mode.

Fix a bug that would set max_tokens automatically even when OPTMAX_NILL was set.

Some code adjustment and maintenance.

---