Easily configure options with new tui form. Bug fixes and code maintenance.
Add runtime configuration form. This is the fastest way to config the script during runtime!
<https://github.com/gptme/gptme/issues/591>

Add chat command `/unkill [NUM]` to uncomment n last entries from the history file.

Set the title of the terminal/multiplexer the script is run to make it easier to find.

Fix bug introduced some days ago which would disrupt the script to have voice in and voice out chats (already comitted).

Improve a hack to retry with curl on w3m gzip error (gzip: stdin: not in gzip format).

Improve avoiding recording internal commands in the user bash history file.

In the session system,
add a "default" option to refer to chatgpt.tsv, which is the default script history file.

Some code maintenance and minor enhancements and bug fixes.

---