Lift input-length limits of some functions. Remove --tmp command line option.
The --tmp option was rather a test code than really good code, so we will just remove this feature as the code does not look like it is the real solution for the problem it is trying to solve!

Maintenance in the function is_txturl().

Improve the syntax of using a subshell in two functions.

Lift the limits and constraints of some functions that were slow before but now can be freed to work with bigger input without too much speed overhead.

Code maintenance.

---