Add magistral model capacity config. Fix token count recording of answers from deepseek models.
------

Auto set model capacity of Magistral models now works correctly.

------

Do not deduce thinking tokens from total when thinking text is available.
This is a fix for mainly deepseek models, cannot test Anthropic api right now and claude-4.7 thinking mode, so token count may be wrongly recorded to the history file when running with anthropic api.

----

Add commented-out code to recognise and exit when user presses CTRL-D in an empty prompt. This mimics bash shell ctrl-d signal handling.

However, I am not sure if we really want to change the current ctrl-d signal behavior in the script which is pretty much the same as an enter key press in an empty prompt.

To exit the script, we currently use CTRL-C.

Ctrl+D should both interrupt and close gptme
https://github.com/gptme/gptme/issues/593

---