Fix code from latest commit to work exactly as before with the '!cat' command.
Fix a little change of behaviour with our last update converning the line:

    var=$(trap "-" INT; bash --norc --noprofile ${@:+-c} "${@}" </dev/tty | tee >(cat >&2) );

As we dont use '$STDERR' here any longer, the only way to disable tee stderr feedback hack is to declare tee as cat temporarily.

----

Re-add but comment out the unused function code removed in the last commit.

---