Update rcmdf() function a nd warn user when a command is blocked.
Also:

Add chat command "!clot" to clot the screen with garbage. It calls the new  function hellof().

I wanted to print the name of the script on start-up as gemini-cli, and claude-code but this  function is too colourful for the sobriety of the start-up chat screen, so lets use this as clotting function.

---