Add chat command combo fature to run single-letter options/commands within chat mode.
----

Add a little code snippet to try and run combo options, such as -CP (Session continue and then Print Session).

I am not sure if there are many useful single-letter command combinations that are worth it because not many options are simple toggles, but this feature is now implemented.

----

Add option so that in chat mode typing '-zw' will turn whisper on and tts on.
This is better than having to type -z and then -w individually.

----

Optimise the cmdf() last clause.

Just compare the first 32 input chars in the fallback shell cmd exec options which was recently added as a feature.

----

DEBUGGING AND INTERNAL FUNCTIONS

Improve command line "option -EE" which must exit the script immediately or asap after executing any command or routine.

The added code will just make the script works as though it was called with "options -EEvvv".

---