Minor improve feature of single-letter combo commands.
Improve how the feature processes many single letter commands or options as a combo option, even with simple numerical (float) arguments, so that even command "-C-M23666/4000za.3" is processed correctly, although really this is just a minor feature and I wish to bother with it no more!

Also, make sure this feature is discreet and to not bother users unnecessarily with error messages.

---