Code maintenance. Optimisation: remove colors from pretty-print response function.
Removing jq syntax to process and display the main text response with colours will improve jq speed itself and most importantly, the in-house foldf() function to fold TTY text. There were too many colour codes when processing a streaming response, and these had to be discarded by foldf(). Now, foldf() will work faster with streaming responses as there is less of a hassle of colour codes. May not be exactly noticeable by the user.

Colour codes for that function would be disabled when markdown processing was enabled, anyways. So better to have response colours solely controlled by the markdown processor software, if the user really wants some colour splash, after all.

--

Update the logging funtion prompt_pf() to incluse some resoning fields that would otherwise only be displayed to the user but would not be logged. We have read that it is expected that the LLM would see its previous thoughts in a new input.

--

Add the "minimal" word in the case statements which set reason effort from the model name, such as `gpt-5-minimal`, now works.

---