Revert last optimisation side effects.
We can safely print a single colour code and then reset colours in the promptf() function without interfering with foldf() beacause now the colour code sequence is not emitted by jq anymore, rather by simple printf commands.

In fact, colourising the response text this way improves how the response was printed when background colour was set, such as in colour theme "Theme 2.1 - Radium Tides".

Se we gained speed with the last code optimisation and now reverted the side effect which was to not print colour codes any longer!

It now even works perfectly with the responses api, so we should see the difference when using it (command line -b or -bb multi-turn chat options).

---