Some maintenance and feature fixes.
Fix thet previous "buggy fix" that was in order to give teh user a chance to abort from the text editor menu on empty output. Now it goes back to readline editor on text editor empty prompt.

Really fix the chat command "/effort" to toggle this parameter off when no input for the command is set.