Adjust and balance valid input sizes to record to the bash history file.
Adjust down $HISTSIZE to 256 (which is user-configurable when using a config file for the script!),

And adjust shell_histf() to accept inputs as long as 16384 chars.

	input: 16384 chars  ~4726 tokens
	max file size: (256*(2^14))/(1024^2) = 4 Mb  ~1,210,000 tokens

Other possible configuration would be:

	alt: $HISTFILE=128, and accept up to 32768 chars in input.

This is extremely good because we shall make better use  of the shell history file!

"Make This Script Bash History Useful Again!" moto effort!