A little more polish to the custom_prf() that was fixed in previous important update.
# Custom prompts
function custom_prf()

Add some more code to return from custom_prf() function when an error occurs,
and minimise the impacts of this error on UX.

---

To remember, last update on this function fixed the following bug:

Fix (old) bug in custom_prf() when returning early from this function.
The bug was activated only in specific cases.

The script generates lots of cache files while in operation.
When multiple instances of chatgpt.sh are running simultaneously,
even when using different history files,
sharing of buffers may lead to some instances of chatgpt.sh to fail.

Although we did ou best to try and minimise this from happening,
coincident multiple curl calls of the script at the same time will corrupt
the buffer files of one instance or even both instances of chatgpt.sh.

The only way to prevent this is setting different $CACHEDIR envars with
dir paths
for multiple running instances of chatgpt.sh.

However, command line --tmp creates a temporary cache dir that is unique
to the instance, usually at /tmp. This feaure also works while in the middle
of a conversation and type in "/tmp" or "!tmp".