Fix regression of the whisper function in chat mode, which would do unnecessary requests on specific situations.
The regression that was fixed in the whisper function in chat mode prevents unnecessary new requests of the same voice input on some specific situations.

Fix a bad and old syntax for the set up of a custom mdf() (markdown function) which any argumens would cause it to fail.

Fix setting curl -F options. A semicolon in this case would cause curl to interpret one argment to -F as multiple ones.