Fix some interface and UX bugs & recent regressions....
Fix a regression with a recent update which tried to fix a problem with printing out a session context snippet on -C Continue option, and now when using the -P to print a whole session to stdout, which became problematic with the just-mentioned bad update that tried to fix context snippet printout on session continuation.

Refix a recent issue that became more problematic in the last few weeks since we detected a bug in the Whisper + Chat mode, specially when when Appending input to the current user prompt with Whisper, or interrupting curl request in that mode, but also when interrupting the curl request in some special/less used cases..

Fix some bugs related to Whisper + Chat input handling.

When on chat mode, interrupting curl request would resubmit the request immediately when option less verbose -v was set. Now fixed.