Fix OpenAI+Groq/Mistral TTS/STT integrations
It turns out that the script was unsetting Mistral and Groq user api keys, so TTS or STT when using different providers in chat mode + speech in + speech out was not working (different service providers).

This bug must have been around for a long time... I remember testing this feature back some ~8 months but apparently I tested only a specific case when the script was run... Such a powerful feature, too bad was broken, but it is now fixed.

Prevent "prompt" property from being set when gpt-4o-transcribe-diarize is used.

Add a little eye-candy for `chatgpt.sh --version`. The eye candy is only displayed when &1 is free (not being piped), which should not disrupt current usage in my git publishing scripts for this project...