Maintenance update.
Improve the sedhtmlf() function to use real html unescaping programmes instead of using a hack. This function is still useful to get the youtube description (and some comments) from a youtube page.

Add some aliases to chat command `/last` and `/last |` to print out the last received message (without or with less).

Add some glob exceptions when automatically selecting whether a model is reasoning or not.

Try to pretty print a bad answer form the server to the use... Using jq instead of cat for that, before the last cat fallback.

Adjust when to sleep on bad response, and give user 4 seconds to check out the bad response.