Update code to work with the new awesome-prompts csv file.
Awesome-prompts recently updated the format of their csv file, so we need to adjust the code to work with it. The new code should also work with older versions of the awesome-prompts csv file.

Update the _dialog_optf(), optimise it and abort on too long loops.

Before names with white spaces, would result in a very long loop in _dialog_optf(), and hang the dialog command.

Improve Awesome UI. And exit on Ctr-C.